pyproject.toml
[project]
name = "form2chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "browser-use>=0.9.5",
    "deepl>=1.24.0",
    "elevenlabs>=2.22.0",
    "google-genai>=1.47.0",
    "markdownify>=1.2.0",
    "openai>=1.109.1",
    "playwright>=1.55.0",
    "python-dotenv>=1.2.1",
    "pyttsx3>=2.99",
    "requests>=2.32.5",
    "textual>=6.5.0",
]
